chore: moved tests from literal_colon_test.go into gin_test.go
